feat: remove duplicate code
